Fixed a bug causing AddVertice to panic due to sending on closed channel
The bug was cause by using GetWait in util with total number of element in the list, then at the
IterChan action, the list has been changed, causing waitgroup to complete early, hence the channel
gets closed. Then there is stll element in IterChan to send to the closed channel.